Add batched VLM parity tests
